+++ packages/framer-motion/src/render/dom/scroll/utils/can-use-native-timeline.ts
@@ -1,0 +1,7 @@
+import { supportsScrollTimeline } from "motion-dom"
+
+export function canUseNativeTimeline(target?: Element) {
+    return (
+        typeof window !== "undefined" && !target && supportsScrollTimeline()
+    )
+}